requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
websockets==15.0.1
gunicorn==23.0.0
python-multipart==0.0.20
pandas==2.3.2
numpy==2.3.2
pydantic-settings==2.10.1
openai==1.99.9
requests==2.32.5
faiss-cpu==1.12.0
sentence-transformers==5.1.0
PyJWT==2.10.1
